Break remaining middleware into modules
